Fixing New Palette error state.
