+++ app/src/main/java/com/akashk/palette/createnewpalette/CreateNewPaletteViewModel.kt
@@ -1,6 +1,7 @@
 package com.akashk.palette.createnewpalette
 
 import androidx.lifecycle.ViewModel
+import com.akashk.palette.R
 import com.akashk.palette.core.ui.UIText
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -11,30 +12,26 @@ import javax.inject.Inject
 class CreateNewPaletteViewModel @Inject constructor() : ViewModel() {
 
     private val _viewState: MutableStateFlow<NewPaletteState> =
-        MutableStateFlow(NewPaletteState.Initial)
+        MutableStateFlow(NewPaletteState())
     val viewState: StateFlow<NewPaletteState> = _viewState
 
     fun enterPaletteName(name: String) {
-        val currentName = viewState.value.paletteName
-        val error = (_viewState.value as? NewPaletteState.Active)?.paletteNameError
-
-        _viewState.value = NewPaletteState.Active(
+        _viewState.value = viewState.value.copy(
             paletteName = name,
             paletteNameError = null
         )
     }
 
-    fun onContinue(name: String) {
+    fun onContinue(name: String, navigateIfValid: () -> Unit) {
         val currentName = viewState.value.paletteName
         if (currentName.isEmpty()) {
-            _viewState.value = NewPaletteState.Active(
+            _viewState.value = viewState.value.copy(
                 paletteName = name,
-                paletteNameError = UIText.StringText("Enter valid palette name")
+                paletteNameError = UIText.ResourceText(R.string.err_enter_valid_palette_name)
             )
             return
+        } else {
+            navigateIfValid.invoke()
         }
-        _viewState.value = NewPaletteState.NavigateToColorPicker(
-            paletteName = currentName
-        )
     }
 }
