+++ app/src/main/res/values-night/strings.xml
@@ -1,3 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">Palette</string>
     <string name="err_enter_valid_palette_name">Enter valid palette name</string>
